fix: split up validation function
