+++ src/bb-functions-publish.ts
@@ -1,17 +1,12 @@
-import chalk from 'chalk';
 import { Command } from 'commander';
 import fs from 'fs-extra';
 import path from 'path';
 import prompts from 'prompts';
 
-import Config from './functions/config';
 import { getAllWasmFunctionsWithVersions } from './functions/functionDefinitions';
 import publishAppFunctions from './functions/publishAppFunctions';
 import { publishWasmBlockStoreFunctions } from './functions/publishWasmBlockStoreFunctions';
-import {
-  FunctionValidator,
-  logValidationResult,
-} from './functions/validations';
+import { validateFunctions } from './functions/validateFunctions';
 
 const program = new Command();
 
@@ -30,44 +25,15 @@ const workingDir = process.cwd();
 
 const baseFunctionsPath = path.join(workingDir, 'functions');
 
-const config = new Config();
-
 const isWasmFunctionProject = fs.existsSync(
   path.join(workingDir, '.wasm-functions'),
 );
 
-const validateFunctions = async (
-  isWasmFunctionProject: boolean,
-): Promise<{ valid: boolean }> => {
-  const validator = new FunctionValidator(config, baseFunctionsPath);
-  await validator.initSchema();
-
-  console.log(chalk.bold(`Validating functions in ${baseFunctionsPath}`));
-
-  const results = await validator.validateFunctions({ isWasmFunctionProject });
-  results.forEach(logValidationResult);
-
-  const valid = results.every((result) => result.status === 'ok');
-
-  if (valid) {
-    console.log(
-      `\n${chalk.green.underline(
-        `✔ All your functions are valid and ready to be published!`,
-      )}`,
-    );
-  } else {
-    console.log(
-      `\n${chalk.red.underline(
-        `✖ Certain functions in your project are invalid.`,
-      )}`,
-    );
-  }
-
-  return { valid };
-};
-
 void (async (): Promise<void> => {
-  const { valid } = await validateFunctions(isWasmFunctionProject);
+  const { valid } = await validateFunctions(
+    isWasmFunctionProject,
+    baseFunctionsPath,
+  );
 
   if (!valid) {
     process.exit(1);
